third_party/blink/web_tests/external/wpt/css/css-ui/tentative/webkit-user-modify-01-expected.txt

This is a testharness.js-based test.
[FAIL] -webkit-user-modify is not supported
  assert_equals: expected (undefined) undefined but got (string) ""
Harness: the test ran to completion.
